Combine library classes into single MTMessage file
